Implements efficient routing for pipeline commands with Redis Cluster. For every command in the pipeline, determines which node is responsible for it and builds a pipeline per node. Each pipeline is first sent to its respective node, then all responses are read back. Any commands that can be retried are done individually.

This leads to a huge performance gains with a pipeline of 100 commands.
Before:
```
time:   [8.3241 ms 8.3772 ms 8.4356 ms]
thrpt:  [11.854 Kelem/s 11.937 Kelem/s 12.013 Kelem/s]
```

Now:
```
time:   [383.37 us 389.66 us 396.19 us]
thrpt:  [252.41 Kelem/s 256.63 Kelem/s 260.84 Kelem/s]
```