@doyshinda 2020-04-06
The clippy failure seems unrelated to my changes:
```
error: called `.as_ref().map(Deref::deref)` (or with one of deref aliases) on an Option value. This can be done more directly by calling `env::args().nth(1).as_deref()` instead
  --> examples/async-connection-loss.rs:61:22
   |
61 |     let mode = match env::args().nth(1).as_ref().map(String::as_str) {
   |                      ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^ help: try using as_deref instead: `env::args().nth(1).as_deref()`
   |
   = note: `-D clippy::option-as-ref-deref` implied by `-D clippy::all`
```