@Terkwood 2020-06-05
re: the clippy failure, you may want to merge master into your branch, as there have been recent updates to master.  It might go away.