@doyshinda 2021-01-09
Hi @Marwes,

I apologize for pinging you directly, but this PR has been open for quite a long time (I forgot about it haha) and it seems you have been active with approving PRs recently.

Looking back, I realize it's very large change set, so I'm going to work on breaking this down in smaller PRs (started with #433 to refactor existing Pipeline logic into it's own file, which is a large part of this change).

My question is: Is there any chance this feature would get some traction? I'm happy to close this PR if not, but either way, it might be nice to get a resolution on this feature.