@rohitjoshi 2021-01-09
Thanks. I have been waiting for pipeline support for cluster connections. Appreciate the work and looking forward for merge