@doyshinda 2021-02-17
Hey @Marwes, I've implemented those changes you suggested, feel free to merge at your earliest convenience. Thanks for taking the time to review this feature!